+++ src/main/java/io/github/torand/openapi2java/collectors/TypeInfoCollector.java
@@ -30,9 +30,13 @@ import static io.github.torand.javacommons.collection.CollectionHelper.nonEmpty;
 import static io.github.torand.javacommons.lang.Exceptions.illegalStateException;
 import static io.github.torand.javacommons.lang.StringHelper.nonBlank;
 import static io.github.torand.javacommons.stream.StreamHelper.streamSafely;
-import static io.github.torand.openapi2java.collectors.Extensions.*;
+import static io.github.torand.openapi2java.collectors.Extensions.EXT_JSON_SERIALIZER;
+import static io.github.torand.openapi2java.collectors.Extensions.EXT_NULLABLE;
+import static io.github.torand.openapi2java.collectors.Extensions.EXT_VALIDATION_CONSTRAINT;
+import static io.github.torand.openapi2java.collectors.Extensions.extensions;
 import static io.github.torand.openapi2java.collectors.TypeInfoCollector.NullabilityResolution.FORCE_NOT_NULLABLE;
 import static io.github.torand.openapi2java.collectors.TypeInfoCollector.NullabilityResolution.FORCE_NULLABLE;
+import static io.github.torand.openapi2java.utils.StringUtils.getClassNameFromFqn;
 import static io.github.torand.openapi2java.utils.StringUtils.joinCsv;
 import static java.lang.Boolean.TRUE;
 import static java.util.Objects.nonNull;
@@ -377,14 +381,6 @@ public class TypeInfoCollector extends BaseCollector {
         }
 
         return typeInfo;
-    }
-
-    private String getClassNameFromFqn(String fqn) {
-        int lastDot = fqn.lastIndexOf(".");
-        if (lastDot == -1) {
-            throw new IllegalStateException("Unexpected fully qualified class name: %s".formatted(fqn));
-        }
-        return fqn.substring(lastDot+1);
     }
 
     private boolean isNullable(Schema<?> schema, NullabilityResolution resolution) {