+++ src/test/java/io/github/torand/openapi2java/ResourceGeneratorTest.java
@@ -21,6 +21,7 @@ import io.swagger.v3.oas.models.OpenAPI;
 import org.junit.jupiter.api.Test;
 
 import static io.github.torand.openapi2java.TestHelper.ConfigVariant.COMMON_HEADERS_FACTORY;
+import static io.github.torand.openapi2java.TestHelper.ConfigVariant.COMMON_PROVIDERS;
 import static io.github.torand.openapi2java.TestHelper.ConfigVariant.OIDC_CLIENT_ANNOTATION;
 import static io.github.torand.openapi2java.TestHelper.ConfigVariant.RESTEASY;
 import static io.github.torand.openapi2java.TestHelper.assertMatchingJavaFiles;
@@ -33,6 +34,7 @@ import static io.github.torand.openapi2java.TestHelper.loadOpenApi30Spec;
 import static io.github.torand.openapi2java.TestHelper.loadOpenApi31Spec;
 import static io.github.torand.openapi2java.TestHelper.withHeadersFactoryOverride;
 import static io.github.torand.openapi2java.TestHelper.withOidcClientAnnotation;
+import static io.github.torand.openapi2java.TestHelper.withProvidersOverride;
 import static io.github.torand.openapi2java.TestHelper.withResteasyResponse;
 import static java.util.Collections.emptyList;
 
@@ -106,6 +108,16 @@ class ResourceGeneratorTest {
     }
 
     @Test
+    void shouldGenerateJavaResource_withProvidersOverride() {
+        Options opts = withProvidersOverride(getJavaOptions());
+        OpenAPI openApiDoc = loadOpenApi31Spec();
+
+        new ResourceGenerator(opts).generate(openApiDoc);
+
+        assertMatchingJavaFilesVariant("%sApi".formatted("Orders"), COMMON_PROVIDERS);
+    }
+
+    @Test
     void shouldGenerateJavaResource_withOidcClientAnnotation() {
         Options opts = withOidcClientAnnotation(getJavaOptions());
         OpenAPI openApiDoc = loadOpenApi31Spec();
@@ -174,6 +186,16 @@ class ResourceGeneratorTest {
         for (String resource : RESOURCES) {
             assertMatchingKotlinFilesVariant("%sApi".formatted(resource), COMMON_HEADERS_FACTORY);
         }
+    }
+
+    @Test
+    void shouldGenerateKotlinResource_withProvidersOverride() {
+        Options opts = withProvidersOverride(getKotlinOptions());
+        OpenAPI openApiDoc = loadOpenApi31Spec();
+
+        new ResourceGenerator(opts).generate(openApiDoc);
+
+        assertMatchingKotlinFilesVariant("%sApi".formatted("Orders"), COMMON_PROVIDERS);
     }
 
     @Test