+++ src/test/resources/expected-output/java/ProductsApi_CommonHeadersFactory.java
@@ -1,5 +1,6 @@
 package io.github.torand.openapi2java.test;
 
+import io.github.torand.openapi2java.test.CommonClientHeadersFactory;
 import io.github.torand.openapi2java.test.model.ProductV1Dto;
 import io.github.torand.openapi2java.test.model.common.ErrorDto;
 import jakarta.validation.Valid;
@@ -47,7 +48,7 @@ import static org.eclipse.microprofile.openapi.annotations.enums.SchemaType.ARRA
 @SecurityRequirement(name = "oidc")
 @Tag(name = "Products", description = "Retrieving and modifying products")
 @RegisterRestClient(configKey = "products-api")
-@RegisterClientHeaders(io.github.torand.openapi2java.test.CommonClientHeadersFactory.class)
+@RegisterClientHeaders(CommonClientHeadersFactory.class)
 @Path(ROOT_PATH)
 public interface ProductsApi_CommonHeadersFactory {