+++ src/test/resources/expected-output/kotlin/OrdersApi_CommonHeadersFactory.kt
@@ -1,5 +1,7 @@
 package io.github.torand.openapi2java.test
 
+import io.github.torand.openapi2java.test.CommonClientHeadersFactory
+import io.github.torand.openapi2java.test.MyClientRequestFilter
 import io.github.torand.openapi2java.test.OrdersApi_CommonHeadersFactory.Companion.ROOT_PATH
 import io.github.torand.openapi2java.test.model.OrderItemV1Dto
 import io.github.torand.openapi2java.test.model.OrderV1Dto
@@ -32,13 +34,15 @@ import org.eclipse.microprofile.openapi.annotations.security.SecurityRequirement
 import org.eclipse.microprofile.openapi.annotations.tags.Tag
 import org.eclipse.microprofile.rest.client.annotation.ClientHeaderParam
 import org.eclipse.microprofile.rest.client.annotation.RegisterClientHeaders
+import org.eclipse.microprofile.rest.client.annotation.RegisterProvider
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient
 
 @SecurityRequirement(name = "oidc")
 @Tag(name = "Orders", description = "Retrieving and modifying orders")
 @RegisterRestClient(configKey = "order-api")
 @ClientHeaderParam(name = AUTHORIZATION, value = [ "Bearer {order-api/mp-rest/api-key}" ])
-@RegisterClientHeaders(io.github.torand.openapi2java.test.CommonClientHeadersFactory::class)
+@RegisterClientHeaders(CommonClientHeadersFactory::class)
+@RegisterProvider(MyClientRequestFilter::class)
 @Path(ROOT_PATH)
 interface OrdersApi_CommonHeadersFactory {